small refactoring for compaction (#27385)
