+++ ydb/core/tx/columnshard/engines/storage/optimizer/lcbuckets/planner/optimizer.cpp
@@ -1,6 +1,6 @@
 #include "optimizer.h"
 
-#include "level/common_level.h"
+#include "level/one_layer.h"
 #include "level/zero_level.h"
 #include "selector/snapshot.h"
 #include "selector/transparent.h"
